Tweaks

+++ Assets/Custom Lighting.hlsl
@@ -10,8 +10,8 @@ void MainLight_float(float3 WorldPos, out float3 Direction, out float3 Color, ou
     ShadowAtten = 1;
 #else
 #if SHADOWS_SCREEN
-    float4 clipPos = TransformWorldToHClip(WorldPos);
-    float4 shadowCoord = ComputeScreenPos(clipPos);
+        float4 clipPos = TransformWorldToHClip(WorldPos);
+        float4 shadowCoord = ComputeScreenPos(clipPos);
 #else
     float4 shadowCoord = TransformWorldToShadowCoord(WorldPos);
 #endif
@@ -32,8 +32,8 @@ void MainLight_half(float3 WorldPos, out half3 Direction, out half3 Color, out h
     ShadowAtten = 1;
 #else
 #if SHADOWS_SCREEN
-    half4 clipPos = TransformWorldToHClip(WorldPos);
-    half4 shadowCoord = ComputeScreenPos(clipPos);
+        half4 clipPos = TransformWorldToHClip(WorldPos);
+        half4 shadowCoord = ComputeScreenPos(clipPos);
 #else
     half4 shadowCoord = TransformWorldToShadowCoord(WorldPos);
 #endif
@@ -65,7 +65,7 @@ void DirectSpecular_half(half3 Specular, half Smoothness, half3 Direction, half3
     Smoothness = exp2(10 * Smoothness + 1);
     WorldNormal = normalize(WorldNormal);
     WorldView = SafeNormalize(WorldView);
-    Out = LightingSpecular(Color, Direction, WorldNormal, WorldView,half4(Specular, 0), Smoothness);
+    Out = LightingSpecular(Color, Direction, WorldNormal, WorldView, half4(Specular, 0), Smoothness);
 #endif
 }
 
@@ -82,7 +82,7 @@ void AdditionalLights_float(float3 SpecColor, float Smoothness, float3 WorldPosi
     for (int i = 0; i < pixelLightCount; ++i)
     {
         Light light = GetAdditionalLight(i, WorldPosition);
-        half3 attenuatedLightColor = light.color * (light.distanceAttenuation * light.shadowAttenuation);
+        float3 attenuatedLightColor = light.color * (light.distanceAttenuation * light.shadowAttenuation);
         diffuseColor += LightingLambert(attenuatedLightColor, light.direction, WorldNormal);
         specularColor += LightingSpecular(attenuatedLightColor, light.direction, WorldNormal, WorldView, float4(SpecColor, 0), Smoothness);
     }
